+++ lib/ui/groups/GroupDetailPanel.dart
@@ -20,6 +20,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_barcode_scanner/flutter_barcode_scanner.dart';
 import 'package:illinois/service/AppDateTime.dart';
 import 'package:illinois/service/FlexUI.dart';
+import 'package:illinois/ui/groups/GroupMemberNotificationsPanel.dart';
 import 'package:illinois/ui/groups/GroupPostDetailPanel.dart';
 import 'package:illinois/ui/widgets/InfoPopup.dart';
 import 'package:rokwire_plugin/model/event.dart';
@@ -110,6 +111,7 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> implements Notifica
   bool               _pollsLoading = false;
 
   bool               _memberAttendLoading = false;
+  bool               _researchProjectConsent = false;
 
   String?            _postId;
 
@@ -179,7 +181,7 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> implements Notifica
   }
 
   bool get _isResearchProject {
-    return (_group?.researchGroup == true);
+    return (_group?.researchProject == true);
   }
 
   bool get _isAttendanceGroup {
@@ -646,21 +648,24 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> implements Notifica
     String members;
     int membersCount = _groupStats?.activeMembersCount ?? 0;
     if (membersCount == 0) {
-      members = Localization().getStringEx("panel.group_detail.members.count.empty", "No Current Members");
+      members = _isResearchProject ? "No Current Participants" : Localization().getStringEx("panel.group_detail.members.count.empty", "No Current Members");
     }
     else if (membersCount == 1) {
-      members = Localization().getStringEx("panel.group_detail.members.count.one", "1 Current Member");
+      members = _isResearchProject ? "1 Current Participant" : Localization().getStringEx("panel.group_detail.members.count.one", "1 Current Member");
     }
     else {
-      members = sprintf(Localization().getStringEx("panel.group_detail.members.count.format", "%s Current Members"),[membersCount]);
+      members = sprintf(_isResearchProject ? "%s Current Participants" : Localization().getStringEx("panel.group_detail.members.count.format", "%s Current Members"), [membersCount]);
     }
 
     int pendingCount = _groupStats?.pendingCount ?? 0;
     String pendingMembers;
     if (_group!.currentUserIsAdmin && pendingCount > 0) {
-      pendingMembers = pendingCount > 1 ?
-        sprintf(Localization().getStringEx("panel.group_detail.pending_members.count.format", "%s Pending Members"), [pendingCount]) :
-        Localization().getStringEx("panel.group_detail.pending_members.count.one", "1 Pending Member");
+      if (pendingCount > 1) {
+        pendingMembers = sprintf(_isResearchProject ? "%s Pending Participants" : Localization().getStringEx("panel.group_detail.pending_members.count.format", "%s Pending Members"), [pendingCount]);
+      }
+      else {
+        pendingMembers = _isResearchProject ? "1 Pending Participant" : Localization().getStringEx("panel.group_detail.pending_members.count.one", "1 Pending Member");
+      }
     }
     else {
       pendingMembers = "";
@@ -682,15 +687,15 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> implements Notifica
     if (_isMemberOrAdmin) {
       if(_isAdmin) {
         commands.add(RibbonButton(
-          label: Localization().getStringEx("panel.group_detail.button.manage_members.title", "Manage Members"),
-          hint: Localization().getStringEx("panel.group_detail.button.manage_members.hint", ""),
+          label: _isResearchProject ? 'Manage Participants' : Localization().getStringEx("panel.group_detail.button.manage_members.title", "Manage Members"),
+          hint: _isResearchProject ? '' : Localization().getStringEx("panel.group_detail.button.manage_members.hint", ""),
           leftIconAsset: 'images/icon-member.png',
           padding: EdgeInsets.symmetric(vertical: 14, horizontal: 0),
           onTap: _onTapMembers,
         ));
         commands.add(Container(height: 1, color: Styles().colors!.surfaceAccent,));
         commands.add(RibbonButton(
-          label: _isResearchProject ? 'Project Settings' : Localization().getStringEx("panel.group_detail.button.group_settings.title", "Group Settings"),
+          label: _isResearchProject ? 'Research Project Settings' : Localization().getStringEx("panel.group_detail.button.group_settings.title", "Group Settings"),
           hint: _isResearchProject ? '' : Localization().getStringEx("panel.group_detail.button.group_settings.hint", ""),
           leftIconAsset: 'images/icon-gear.png',
           padding: EdgeInsets.symmetric(vertical: 14, horizontal: 0),
@@ -720,6 +725,17 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> implements Notifica
           ]));
         }
       }
+      if (CollectionUtils.isNotEmpty(commands)) {
+        commands.add(Container(height: 1, color: Styles().colors!.surfaceAccent));
+      }
+      commands.add(RibbonButton(
+        label: Localization().getStringEx("panel.group_detail.button.notifications.title", "Notifications"),
+        hint: Localization().getStringEx("panel.group_detail.button.notifications.hint", ""),
+        leftIconAsset: 'images/icon-reminder.png',
+        leftIconPadding: EdgeInsets.only(right: 8, left: 2),
+        padding: EdgeInsets.symmetric(vertical: 14),
+        onTap: _onTapNotifications,
+      ));
       if (StringUtils.isNotEmpty(_group?.webURL)) {
         commands.add(Container(height: 1, color: Styles().colors!.surfaceAccent));
         commands.add(_buildWebsiteLink());
@@ -1055,8 +1071,8 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> implements Notifica
   }
 
   Widget _buildBadgeOrCategoryWidget() {
-    return _showMembershipBadge ?
-      Row(children: <Widget>[
+    if (_showMembershipBadge) {
+      return Row(children: <Widget>[
         Container(padding: EdgeInsets.symmetric(horizontal: 8, vertical: 4), decoration: BoxDecoration(color: _group!.currentUserStatusColor, borderRadius: BorderRadius.all(Radius.circular(2)),), child:
           Center(child:
             Semantics(label: _group?.currentUserStatusText?.toLowerCase(), excludeSemantics: true, child:
@@ -1066,14 +1082,16 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> implements Notifica
         ),
         Expanded(child: Container(),),
         _buildPolicyButton(),
-      ],) :
-    
-      Row(children: <Widget>[
+      ],);
+    }
+    else {
+      return Row(children: <Widget>[
         Expanded(child:
-          Text(_group?.category?.toUpperCase() ?? '', style:  Styles().textStyles?.getTextStyle('widget.title.tiny'),),
+          Text(_isResearchProject ? '' : (_group?.category?.toUpperCase() ?? ''), style:  Styles().textStyles?.getTextStyle('widget.title.tiny'),),
         ),
         _buildPolicyButton(),
       ],);
+    }
   }
 
   Widget _buildPolicyButton() {
@@ -1122,48 +1140,75 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> implements Notifica
   }
 
   Widget _buildMembershipRequest() {
-    return
-      Auth2().isOidcLoggedIn && _group!.currentUserCanJoin
-          ? Container(color: Colors.white,
-              child: Padding(padding: EdgeInsets.all(16),
-                  child: RoundedButton(label: Localization().getStringEx("panel.group_detail.button.request_to_join.title",  'Request to join'),
-                    backgroundColor: Styles().colors!.white,
-                    textColor: Styles().colors!.fillColorPrimary,
-                    fontFamily: Styles().fontFamilies!.bold,
-                    fontSize: 16,
-                    padding: EdgeInsets.symmetric(horizontal: 32, vertical: 12),
-                    borderColor: Styles().colors!.fillColorSecondary,
-                    borderWidth: 2,
-                    onTap:() { _onMembershipRequest();  }
+    if (Auth2().isOidcLoggedIn && _group!.currentUserCanJoin) {
+      bool showConsent = (_group?.researchProject == true) && (_group?.researchConfirmation == true);
+      bool requestToJoinEnabled = (_group?.researchProject != true) || (_group?.researchConfirmation != true) || _researchProjectConsent;
+      return Container(decoration: BoxDecoration(color: Styles().colors?.white, border: Border(top: BorderSide(color: Styles().colors!.surfaceAccent!, width: 1))), child:
+        Padding(padding: EdgeInsets.zero, child:
+          Column(children: [
+            Visibility(visible: showConsent, child:
+              Row(children: [
+                InkWell(onTap: _onResearchProjectConsent, child:
+                  Padding(padding: EdgeInsets.all(16), child:
+                    Image.asset(_researchProjectConsent ? "images/selected-checkbox.png" : "images/deselected-checkbox.png"),
                   ),
+                ),
+                Expanded(child:
+                  Padding(padding: EdgeInsets.only(right: 16, top: 12, bottom: 12), child:
+                    Text('I consent to participate in this study and have read all the information provided.', style: Styles().textStyles?.getTextStyle("widget.detail.regular"), textAlign: TextAlign.left,)
+                  ),
+                ),
+              ]),
+            ),
+            Padding(padding: EdgeInsets.only(left: 16, right: 16, top: showConsent ? 0 : 16, bottom: 16), child:
+              RoundedButton(label: _isResearchProject ? "Request to participate" : Localization().getStringEx("panel.group_detail.button.request_to_join.title",  'Request to join'),
+                backgroundColor: Styles().colors!.white,
+                textColor: requestToJoinEnabled ? Styles().colors!.fillColorPrimary : Styles().colors!.surfaceAccent,
+                fontFamily: Styles().fontFamilies!.bold,
+                fontSize: 16,
+                padding: EdgeInsets.symmetric(horizontal: 32, vertical: 12),
+                borderColor: requestToJoinEnabled ? Styles().colors!.fillColorSecondary : Styles().colors!.surfaceAccent,
+                borderWidth: 2,
+                onTap:() { _onMembershipRequest();  }
               ),
-            )
-          : Container();
+            ),
+          ],),
+        ),
+      );
+    }
+    else {
+      return Container();
+    }
+  }
+
+  void _onResearchProjectConsent() {
+    setState(() {
+      _researchProjectConsent = !_researchProjectConsent;
+    });
   }
 
   Widget _buildCancelMembershipRequest() {
-    return
-      Auth2().isOidcLoggedIn && _group!.currentUserIsPendingMember
-          ? Stack(
-            alignment: Alignment.center,
-            children: [
-              Container(color: Colors.white,
-                  child: Padding(padding: EdgeInsets.all(16),
-                    child: RoundedButton(label: Localization().getStringEx("panel.group_detail.button.cancel_request.title",  'Cancel Request'),
-                        backgroundColor: Styles().colors!.white,
-                        textColor: Styles().colors!.fillColorPrimary,
-                        fontFamily: Styles().fontFamilies!.bold,
-                        fontSize: 16,
-                        padding: EdgeInsets.symmetric(horizontal: 32, vertical: 12),
-                        borderColor: Styles().colors!.fillColorSecondary,
-                        borderWidth: 2,
-                        onTap:() { _onCancelMembershipRequest();  }
-                    ),
-                  )),
-              _confirmationLoading ? CircularProgressIndicator(strokeWidth: 2, valueColor: AlwaysStoppedAnimation<Color?>(Styles().colors!.fillColorSecondary), ) : Container(),
-            ],
-          )
-          : Container();
+    if (Auth2().isOidcLoggedIn && _group!.currentUserIsPendingMember) {
+      return Container(decoration: BoxDecoration(color: Styles().colors?.white, border: Border(top: BorderSide(color: Styles().colors!.surfaceAccent!, width: 1))), child:
+        Padding(padding: EdgeInsets.all(16), child:
+          RoundedButton(label: Localization().getStringEx("panel.group_detail.button.cancel_request.title",  'Cancel Request'),
+            backgroundColor: Styles().colors!.white,
+            textColor: Styles().colors!.fillColorPrimary,
+            fontFamily: Styles().fontFamilies!.bold,
+            fontSize: 16,
+            padding: EdgeInsets.symmetric(horizontal: 32, vertical: 12),
+            borderColor: Styles().colors!.fillColorSecondary,
+            borderWidth: 2,
+            progress: _confirmationLoading,
+            onTap:() { _onCancelMembershipRequest();  }
+          ),
+        )
+      );
+    }
+    else {
+      return Container();
+    }
+      
   }
 
   Widget _buildConfirmationDialog({String? confirmationTextMsg,
@@ -1265,30 +1310,6 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> implements Notifica
                                   onPositiveTap: _onTapLeaveDialog)).then((value) => Navigator.pop(context));
                         })),
                 Visibility(
-                    visible: _canEditGroup,
-                    child: RibbonButton(
-                        leftIconAsset: "images/icon-gear.png",
-                        label: _isResearchProject ? 'Project Settings' : Localization().getStringEx("panel.group_detail.button.group.edit.title", "Group Settings"),
-                        onTap: () {
-                          Navigator.pop(context);
-                          _onTapSettings();
-                        })),
-                Visibility(
-                    visible: _canDeleteGroup,
-                    child: RibbonButton(
-                        leftIconAsset: "images/icon-delete-group.png",
-                        label: _isResearchProject ? 'Delete project' : Localization().getStringEx("panel.group_detail.button.group.delete.title", "Delete group"),
-                        onTap: () {
-                          Analytics().logSelect(target: "Delete group", attributes: _group?.analyticsAttributes);
-                          showDialog(
-                              context: context,
-                              builder: (context) => _buildConfirmationDialog(
-                                  confirmationTextMsg: confirmMsg,
-                                  positiveButtonLabel: Localization().getStringEx('dialog.yes.title', 'Yes'),
-                                  negativeButtonLabel: Localization().getStringEx('dialog.no.title', 'No'),
-                                  onPositiveTap: _onTapDeleteDialog)).then((value) => Navigator.pop(context));
-                        })),
-                Visibility(
                     visible: _canAddEvent,
                     child: RibbonButton(
                         leftIconAsset: "images/icon-edit.png",
@@ -1305,6 +1326,30 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> implements Notifica
                         onTap: (){
                           Navigator.pop(context);
                           _onTapCreateEvent();
+                        })),
+                Visibility(
+                    visible: _canEditGroup,
+                    child: RibbonButton(
+                        leftIconAsset: "images/icon-gear.png",
+                        label: _isResearchProject ? 'Research project settings' : Localization().getStringEx("panel.group_detail.button.group.edit.title", "Group Settings"),
+                        onTap: () {
+                          Navigator.pop(context);
+                          _onTapSettings();
+                        })),
+                Visibility(
+                    visible: _canDeleteGroup,
+                    child: RibbonButton(
+                        leftIconAsset: "images/icon-delete-group.png",
+                        label: _isResearchProject ? 'Delete research project' : Localization().getStringEx("panel.group_detail.button.group.delete.title", "Delete group"),
+                        onTap: () {
+                          Analytics().logSelect(target: "Delete group", attributes: _group?.analyticsAttributes);
+                          showDialog(
+                              context: context,
+                              builder: (context) => _buildConfirmationDialog(
+                                  confirmationTextMsg: confirmMsg,
+                                  positiveButtonLabel: Localization().getStringEx('dialog.yes.title', 'Yes'),
+                                  negativeButtonLabel: Localization().getStringEx('dialog.no.title', 'No'),
+                                  onPositiveTap: _onTapDeleteDialog)).then((value) => Navigator.pop(context));
                         })),
               ]));
         });
@@ -1431,6 +1476,11 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> implements Notifica
   void _onTapPromote() {
     Analytics().logSelect(target: "Promote Group", attributes: _group?.analyticsAttributes);
     Navigator.push(context, CupertinoPageRoute(builder: (context) => GroupQrCodePanel(group: _group)));
+  }
+
+  void _onTapNotifications() {
+    Analytics().logSelect(target: "Notifications", attributes: _group?.analyticsAttributes);
+    Navigator.push(context, CupertinoPageRoute(builder: (context) => GroupMemberNotificationsPanel(groupId: _group?.id, memberId: _group?.currentMember?.id)));
   }
 
   void _onTapTakeAttendance() {